Feature: multi-macos (#26)

+++ install/prereq-osx.sh
@@ -6,10 +6,13 @@ unamestr=$(uname)
 
 # Mac OS X
 if [[ $unamestr == "Darwin" ]]; then
+
+    if [[ "${SKIP_DOTFILES_FORCE_INSTALL_HOMEBREW}x" == "x" ]]; then
     pip || easy_install pip
     brew || ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
     brew cask || brew tap caskroom/cask
     brew install ansible
+    fi
 # debian, ubuntu, mint etc.
 elif [[ $unamestr == "Linux"  && -f $(which apt-get) ]]; then
     sudo apt-get install -y software-properties-common
